Splitting push manager. Extracting sse connection managment from push manager to new component to handle streaming connection
