+++ src/main/java/io/split/android/client/service/sseclient/SseConnectionManager.java
@@ -1,0 +1,9 @@
+package io.split.android.client.service.sseclient;
+
+import io.split.android.client.lifecycle.SplitLifecycleAware;
+
+public interface SseConnectionManager extends SplitLifecycleAware {
+    void start();
+    void stop();
+    void setListener(SseConnectionManagerListener listener);
+}
