+++ src/main/java/io/split/android/client/service/sseclient/SseConnectionManagerListener.java
@@ -1,0 +1,6 @@
+package io.split.android.client.service.sseclient;
+
+public interface SseConnectionManagerListener {
+    void onSseAvailable();
+    void onSseNotAvailable();
+}
