+++ src/test/java/io/split/android/client/service/sseclient/PushNotificationManagerTest.java
@@ -15,8 +15,6 @@ import java.util.Map;
 
 import io.split.android.client.SplitClientConfig;
 import io.split.android.client.service.executor.SplitTaskExecutionInfo;
-import io.split.android.client.service.executor.SplitTaskExecutor;
-import io.split.android.client.service.executor.SplitTaskFactory;
 import io.split.android.client.service.executor.SplitTaskType;
 import io.split.android.client.service.sseauthentication.SseAuthenticationTask;
 import io.split.android.client.service.sseclient.feedbackchannel.PushManagerEventBroadcaster;
@@ -50,15 +48,12 @@ public class PushNotificationManagerTest {
 
     @Mock
     SseClient mSseClient;
-    @Mock
-    SplitTaskExecutor mTaskExecutor;
-    @Mock
-    SplitTaskFactory mSplitTaskFactory;
+
     @Mock
     PushManagerEventBroadcaster mBroadcasterChannel;
 
     @Mock
-    SseAuthenticationTask mSseAuthTask;
+    SseConnectionManager mSseConnectionManager;
 
     @Mock
     NotificationParser mNotificationParser;
@@ -66,11 +61,6 @@ public class PushNotificationManagerTest {
     @Mock
     NotificationProcessor mNotificationProcessor;
 
-    @Mock
-    ReconnectBackoffCounter mAuthBackoffCounter;
-
-    @Mock
-    ReconnectBackoffCounter mSseBackoffCounter;
 
     PushNotificationManager mPushManager;
 
@@ -78,194 +68,11 @@ public class PushNotificationManagerTest {
     public void setup() {
 
         MockitoAnnotations.initMocks(this);
-        when(mAuthBackoffCounter.getNextRetryTime()).thenReturn(1L);
-        when(mSseBackoffCounter.getNextRetryTime()).thenReturn(1L);
-        mPushManager = new PushNotificationManager(mSseClient, mTaskExecutor,
-                mSplitTaskFactory, mNotificationParser, mNotificationProcessor, mBroadcasterChannel,
-                mAuthBackoffCounter, mSseBackoffCounter);
+
+        mPushManager = new PushNotificationManager(mSseClient,
+                mNotificationParser, mNotificationProcessor, mBroadcasterChannel, mSseConnectionManager);
     }
 
-    @Test
-    public void authOkAndSubscritionToSse() {
-        List<String> channels = dummyChannels();
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-        mPushManager.onOpen();
-
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-        verify(mSseClient, times(1)).connect(TOKEN, channels);
-        verify(mBroadcasterChannel, times(2)).pushMessage(any());
-        verify(mAuthBackoffCounter, times(1)).resetCounter();
-        verify(mSseBackoffCounter, times(1)).resetCounter();
-    }
-
-    @Test
-    public void sseAuthCredentialsError() {
-        List<String> channels = dummyChannels();
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-
-        Map<String, Object> respData = new HashMap<>();
-        respData.put(SplitTaskExecutionInfo.IS_VALID_API_KEY, false);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.error(SplitTaskType.SSE_AUTHENTICATION_TASK, respData));
-
-        ArgumentCaptor<Long> reconnectTime = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-        verify(mTaskExecutor, never()).schedule(any(SseAuthenticationTask.class), reconnectTime.capture(), any(PushNotificationManager.class));
-        verify(mSseClient, never()).connect(any(), any());
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-        verify(mAuthBackoffCounter, never()).getNextRetryTime();
-    }
-
-    @Test
-    public void sseAuthUnexpectedError() {
-        List<String> channels = dummyChannels();
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        Map<String, Object> respData = new HashMap<>();
-        respData.put(SplitTaskExecutionInfo.UNEXPECTED_ERROR, true);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.error(SplitTaskType.SSE_AUTHENTICATION_TASK, respData));
-
-        ArgumentCaptor<Long> reconnectTime = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-        verify(mTaskExecutor, times(1)).schedule(any(SseAuthenticationTask.class), reconnectTime.capture(), any(PushNotificationManager.class));
-        verify(mSseClient, never()).connect(any(), any());
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-        verify(mAuthBackoffCounter, times(1)).getNextRetryTime();
-    }
-
-    @Test
-    public void sseAuthStreamingDisabled() {
-        List<String> channels = dummyChannels();
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        Map<String, Object> respData = new HashMap<>();
-        respData.put(SplitTaskExecutionInfo.IS_VALID_API_KEY, true);
-        respData.put(SplitTaskExecutionInfo.IS_STREAMING_ENABLED, false);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.error(SplitTaskType.SSE_AUTHENTICATION_TASK, respData));
-
-        ArgumentCaptor<Long> reconnectTime = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-        verify(mTaskExecutor, times(1)).schedule(any(SseAuthenticationTask.class), reconnectTime.capture(), any(PushNotificationManager.class));
-        verify(mSseClient, never()).connect(any(), any());
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-        verify(mAuthBackoffCounter, times(1)).getNextRetryTime();
-    }
-
-    @Test
-    public void authOkAndSubscritionToSseRecoverableError() {
-        List<String> channels = dummyChannels();
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-        mPushManager.onError(true);
-
-        verify(mTaskExecutor, times(1)).schedule(
-                any(PushNotificationManager.SseReconnectionTimer.class),
-                anyLong(),
-                isNull());
-        verify(mSseClient, times(1)).connect(TOKEN, channels);
-        ArgumentCaptor<String> keepAliveId = ArgumentCaptor.forClass(String.class);
-        // Fist time connecting should not be called because keepalive timer is not set yet
-        verify(mTaskExecutor, never()).stopTask(anyString());
-
-        verify(mSseBackoffCounter, times(1)).getNextRetryTime();
-
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-
-        verify(mAuthBackoffCounter, times(1)).resetCounter();
-
-    }
-
-    @Test
-    public void recoverableSseErrorWhileConnected() {
-        List<String> channels = dummyChannels();
-        String keepAliveTaskId = "id1";
-        when(mTaskExecutor.schedule(any(PushNotificationManager.SseKeepAliveTimer.class),
-                anyLong(), isNull())).thenReturn(keepAliveTaskId);
-        mPushManager.start();
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-
-        mPushManager.onOpen();
-        // Reset broadcaster channel to count only messages delivered after on error
-        reset(mBroadcasterChannel);
-        reset(mSseBackoffCounter);
-        mPushManager.onError(true);
-
-        verify(mTaskExecutor, times(1)).schedule(
-                any(PushNotificationManager.SseReconnectionTimer.class),
-                anyLong(),
-                isNull());
-        verify(mSseBackoffCounter, times(1)).getNextRetryTime();
-        verify(mSseClient, times(1)).connect(TOKEN, channels);
-        ArgumentCaptor<String> keepAliveId = ArgumentCaptor.forClass(String.class);
-        verify(mTaskExecutor, times(1)).stopTask(keepAliveTaskId);
-
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-
-    }
-
-    @Test
-    public void authOkAndSubscritionToSseUnrecoverableError() {
-        List<String> channels = dummyChannels();
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-        mPushManager.onError(false);
-
-        verify(mTaskExecutor, never()).schedule(
-                any(PushNotificationManager.SseReconnectionTimer.class),
-                anyLong(),
-                any(PushNotificationManager.class));
-        verify(mSseClient, times(1)).connect(TOKEN, channels);
-        ArgumentCaptor<String> keepAliveId = ArgumentCaptor.forClass(String.class);
-        // Fist time connecting should not be called because keepalive timer is not set yet
-        verify(mTaskExecutor, never()).stopTask(anyString());
-
-        verify(mSseBackoffCounter, never()).getNextRetryTime();
-
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-
-        verify(mAuthBackoffCounter, times(1)).resetCounter();
-
-    }
-
-    @Test
-    public void authOkAndChannelsError() {
-        List<String> channels = new ArrayList<>();
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-        verify(mSseClient, never()).connect(TOKEN, channels);
-        ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
-        verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
-        Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-    }
 
     @Test
     public void onMessageToProcess() {
@@ -274,24 +81,13 @@ public class PushNotificationManagerTest {
         when(mNotificationParser.parseIncoming(anyString()))
                 .thenReturn(new IncomingNotification(NotificationType.SPLIT_KILL,
                         "channel", "{}", 1L));
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
+
         mPushManager.start();
-        reset(mTaskExecutor);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
 
         mPushManager.onMessage(message(data));
 
         verify(mNotificationProcessor, times(1))
                 .process(any(IncomingNotification.class));
-        ArgumentCaptor<Long> downNotificationTime = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1))
-                .schedule(any(PushNotificationManager.SseKeepAliveTimer.class),
-                        downNotificationTime.capture(), isNull());
-        verify(mTaskExecutor, times(1))
-                .schedule(any(PushNotificationManager.SseKeepAliveTimer.class),
-                        downNotificationTime.capture(), isNull());
-        Assert.assertEquals(70L, downNotificationTime.getValue().longValue());
     }
 
     @Test
@@ -305,11 +101,8 @@ public class PushNotificationManagerTest {
                         "control_pri", data, 1L));
         when(mNotificationParser.parseOccupancy(anyString()))
                 .thenReturn(occupancyNotification);
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
+
         mPushManager.start();
-        reset(mTaskExecutor);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
 
         // Enable polling prior to disable it
         mPushManager.notifyPollingEnabled();
@@ -319,13 +112,6 @@ public class PushNotificationManagerTest {
         verify(mNotificationProcessor, times(0))
                 .process(any(IncomingNotification.class));
         ArgumentCaptor<Long> downNotificationTime = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1))
-                .schedule(any(PushNotificationManager.SseKeepAliveTimer.class),
-                        downNotificationTime.capture(), isNull());
-        verify(mTaskExecutor, times(1))
-                .schedule(any(PushNotificationManager.SseKeepAliveTimer.class),
-                        downNotificationTime.capture(), isNull());
-        Assert.assertEquals(70L, downNotificationTime.getValue().longValue());
         ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
         verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
         Assert.assertEquals(DISABLE_POLLING, messageCaptor.getValue().getMessage());
@@ -342,11 +128,8 @@ public class PushNotificationManagerTest {
                         "control_sec", data, 1L));
         when(mNotificationParser.parseOccupancy(anyString()))
                 .thenReturn(occupancyNotification);
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
 
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
+        mPushManager.start();
 
         // Enable polling prior to disable it
         mPushManager.notifyPollingEnabled();
@@ -379,23 +162,13 @@ public class PushNotificationManagerTest {
                 .thenReturn(incomingNotification);
         when(mNotificationParser.parseControl(anyString()))
                 .thenReturn(streamingDisabledNotification);
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseKeepAliveTimer.class), anyLong(), any()))
-                .thenReturn(keepAliveTaskId);
+        ;
 
         mPushManager.start();
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseTokenExpiredTimer.class), anyLong(), any()))
-                .thenReturn(refreshTokenTaskId);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-        mPushManager.onOpen();
+        mPushManager.onSseAvailable();
         reset(mBroadcasterChannel);
         mPushManager.onMessage(message(""));
 
-        verify(mTaskExecutor, times(2)).stopTask(keepAliveTaskId);
-        verify(mTaskExecutor, times(1)).stopTask(refreshTokenTaskId);
         ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
         verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
         Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
@@ -408,34 +181,21 @@ public class PushNotificationManagerTest {
                 "control_pri", "", 100L);
         ControlNotification streamingPausedNotification
                 = Json.fromJson("{\"type\":\"CONTROL\",\"controlType\":\"STREAMING_PAUSED\"}", ControlNotification.class);
-        String keepAliveTaskId = "id1";
-        String refreshTokenTaskId = "id2";
         List<String> channels = new ArrayList<>();
         channels.add("dummychannel");
         when(mNotificationParser.parseIncoming(anyString()))
                 .thenReturn(incomingNotification);
         when(mNotificationParser.parseControl(anyString()))
                 .thenReturn(streamingPausedNotification);
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseKeepAliveTimer.class), anyLong(), any()))
-                .thenReturn(keepAliveTaskId);
 
         mPushManager.start();
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseTokenExpiredTimer.class), anyLong(), any()))
-                .thenReturn(refreshTokenTaskId);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-        mPushManager.onOpen();
+        mPushManager.onSseAvailable();
         reset(mBroadcasterChannel);
         mPushManager.onMessage(message(""));
 
-        verify(mTaskExecutor, times(1)).stopTask(anyString());
         ArgumentCaptor<PushStatusEvent> messageCaptor = ArgumentCaptor.forClass(PushStatusEvent.class);
         verify(mBroadcasterChannel, times(1)).pushMessage(messageCaptor.capture());
         Assert.assertEquals(ENABLE_POLLING, messageCaptor.getValue().getMessage());
-
     }
 
     @Test
@@ -455,18 +215,9 @@ public class PushNotificationManagerTest {
                 .thenReturn(incomingNotification);
         when(mNotificationParser.parseControl(anyString()))
                 .thenReturn(streamingDisabledNotification);
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseKeepAliveTimer.class), anyLong(), any()))
-                .thenReturn(keepAliveTaskId);
 
         mPushManager.start();
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseTokenExpiredTimer.class), anyLong(), any()))
-                .thenReturn(refreshTokenTaskId);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-        mPushManager.onOpen();
+        mPushManager.onSseAvailable();
         mPushManager.onMessage(message(data));
 
         reset(mNotificationParser);
@@ -478,160 +229,6 @@ public class PushNotificationManagerTest {
         mPushManager.onMessage(message(data));
 
         verify(mNotificationProcessor, never()).process(any());
-    }
-
-
-    @Test
-    public void onKeepAlive() {
-        List<String> channels = new ArrayList<>();
-        String data = "{}";
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        reset(mTaskExecutor);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true)));
-
-        mPushManager.onKeepAlive();
-
-        ArgumentCaptor<Long> downNotificationTime = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1)).schedule(any(PushNotificationManager.SseKeepAliveTimer.class), downNotificationTime.capture(), any());
-        Assert.assertEquals(70L, downNotificationTime.getValue().longValue());
-    }
-
-    @Test
-    public void setupSseTokenExpirationTimerOnAuth() throws InterruptedException {
-        List<String> channels = dummyChannels();
-        String data = "{}";
-
-        long expirationTime = System.currentTimeMillis() / 1000 + 603;
-
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.start();
-        reset(mTaskExecutor);
-        mPushManager.taskExecuted(SplitTaskExecutionInfo.success(SplitTaskType.SSE_AUTHENTICATION_TASK,
-                buildAuthMap(TOKEN, channels, true, true, expirationTime
-                )));
-
-        verify(mSseClient, times(1)).connect(TOKEN, channels);
-        ArgumentCaptor<Long> expirationCaptor = ArgumentCaptor.forClass(Long.class);
-        verify(mTaskExecutor, times(1)).schedule(any(PushNotificationManager.SseTokenExpiredTimer.class), expirationCaptor.capture(), any());
-        Assert.assertEquals(3, expirationCaptor.getValue().longValue());
-    }
-
-    @Test
-    public void refreshSseToken() throws InterruptedException {
-        String keepAliveTaskId = "keepAliveTaskId";
-        List<String> channels = dummyChannels();
-        String data = "{}";
-        mPushManager.start();
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        when(mTaskExecutor.schedule(
-                any(PushNotificationManager.SseKeepAliveTimer.class), anyLong(), any()))
-                .thenReturn(keepAliveTaskId);
-
-
-        PushNotificationManager.SseTokenExpiredTimer tokenExpiredTimerTask
-                = mPushManager.new SseTokenExpiredTimer();
-
-        mPushManager.onKeepAlive();
-        tokenExpiredTimerTask.execute();
-        verify(mTaskExecutor, times(1)).stopTask(keepAliveTaskId);
-        verify(mSseClient, times(1)).disconnect();
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-    }
-
-    @Test
-    public void disconnectOnBg() throws InterruptedException {
-        mPushManager.start();
-        mPushManager.pause();
-        verify(mSseClient, times(1)).scheduleDisconnection(anyLong());
-    }
-
-    @Test
-    public void connectOnFg() throws InterruptedException {
-        when(mSseClient.cancelDisconnectionTimer()).thenReturn(true);
-        when(mSseClient.readyState()).thenReturn(SseClient.CLOSED);
-        mPushManager.start();
-        mPushManager.pause();
-        mPushManager.resume();
-        verify(mSseClient, times(0)).cancelDisconnectionTimer();
-    }
-
-    @Test
-    public void noConnectOnFgWhileStillConnected() throws InterruptedException {
-        when(mSseClient.cancelDisconnectionTimer()).thenReturn(true);
-        when(mSseClient.readyState()).thenReturn(SseClient.OPEN);
-        mPushManager.start();
-        mPushManager.pause();
-        mPushManager.resume();
-        verify(mSseClient, times(1)).cancelDisconnectionTimer();
-        verify(mTaskExecutor, never()).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-    }
-
-    @Test
-    public void connectOnFgWhileOpenAndCancelFail() throws InterruptedException {
-        // Tests that connection is triggered if ready state is open
-        // but couldn't cancel scheduled disconnection because it was triggered
-        when(mSseClient.cancelDisconnectionTimer()).thenReturn(false);
-        when(mSseClient.readyState()).thenReturn(SseClient.OPEN);
-        mPushManager.start();
-        // Should be here to avoid counting start two
-        when(mSplitTaskFactory.createSseAuthenticationTask()).thenReturn(mSseAuthTask);
-        mPushManager.pause();
-        mPushManager.resume();
-        verify(mSseClient, times(1)).cancelDisconnectionTimer();
-        verify(mTaskExecutor, times(1)).submit(any(SseAuthenticationTask.class), any(PushNotificationManager.class));
-    }
-
-
-    @Test
-    public void sseClientError() throws InterruptedException {
-        when(mSseClient.readyState()).thenReturn(SseClient.OPEN);
-        mPushManager.start();
-        mPushManager.onError(true);
-        verify(mTaskExecutor, times(1)).schedule(any(PushNotificationManager.SseReconnectionTimer.class), anyLong(), isNull());
-    }
-
-    @Test
-    public void sseClientErrorOnBg() throws InterruptedException {
-        when(mSseClient.readyState()).thenReturn(SseClient.OPEN);
-        mPushManager.start();
-        mPushManager.pause();
-        mPushManager.onError(true);
-        verify(mTaskExecutor, never()).schedule(any(PushNotificationManager.SseReconnectionTimer.class), anyLong(), isNull());
-    }
-
-    @After
-    public void teardDown() {
-        reset();
-    }
-
-    private Map<String, Object> buildAuthMap(String token,
-                                             List<String> channels,
-                                             boolean isApiKeyValid,
-                                             boolean isStreamingEnabled) {
-        return buildAuthMap(token, channels, isApiKeyValid, isStreamingEnabled,
-                9999999L);
-    }
-
-
-    private Map<String, Object> buildAuthMap(String token, List<String> channels,
-                                             boolean isApiKeyValid, boolean isStreamingEnabled,
-                                             long expirationTime) {
-        Map<String, Object> data = new HashMap<>();
-        SseJwtToken jwtToken = new SseJwtToken(expirationTime, channels, TOKEN);
-        data.put(SplitTaskExecutionInfo.PARSED_SSE_JWT, jwtToken);
-        data.put(SplitTaskExecutionInfo.IS_VALID_API_KEY, isApiKeyValid);
-        data.put(SplitTaskExecutionInfo.IS_STREAMING_ENABLED, isStreamingEnabled);
-        return data;
-    }
-
-    private List<String> dummyChannels() {
-        List<String> channels = new ArrayList<>();
-        channels.add("channel1");
-        channels.add("channel2");
-        return channels;
     }
 
     private Map<String, String> message(String data) {
